feat: Implement useRoomSelector in Roster component
